requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.6
comm==0.1.3
crashtest==0.4.1
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.3
django-bootstrap4==23.2
django-cors-headers==4.2.0
django-elasticsearch-dsl==7.3
django-extensions==3.2.3
djangorestframework==3.14.0
dulwich==0.21.5
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
html5lib==1.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keyring==23.13.1
langchain==0.0.224
langchainplus-sdk==0.0.20
lockfile==0.12.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.1
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rapidfuzz==2.15.1
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.11
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
sendgrid==6.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starkbank-ecdsa==2.2.0
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trove-classifiers==2023.7.6
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.24.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
whitenoise==6.5.0
widgetsnbextension==4.0.8
Wikipedia-API==0.6.0
xattr==0.10.1
yarl==1.9.2
zipp==3.16.2
